Rough draft of the player positioning and animation system
The "get_tile_coords" function in the "common_definitions.py" module was repurposed to a more generic "get_tile_related_screen_coords" function (in particular, the previously defined "x_offset" is now a parameter of the function, and an ulterior "y_offset" parameter was also defined).

Said function is used to calculate the screen positioning of - potentially - every possible sprite relatively to a 26x15 grid of 24x24 tiles which occupies the entirety of the screen (save for the two 8x360 boxes positioned leftmost and rightmost on the screen, respectively).

[N.B.]:
The player's idle animation gets played correctly (each frame gets substituted by the next, there's no frame overlapping), even without cutting a 36x48 subsurface from the "playing_surface" beneath the player and blitting it in the following game loop before blitting the next frame of the player's idle animation itself.

This may be precisely because of the fact the player's sprite gets blitted on top of the "game_surface", and not on top of the "playing_surface", which remains unchanged at every game loop iteration.

If - when testing the player's movement - the player's sprite doesn't leave a "trail" behind it, we should probably get rid of the subsurface cutting.

[TO DO]:
Cap the framerate at 60FPS: right now, even though a "clock" object was introduced, the game still seems to run at an excessively high uncapped framerate.

This may be the issue with the highest priority so far; thus, the next commit will probably be primarily dedicated to solve it.