Animation speed issue solved + subsurface extraction removed + rough player movement implemented
The game was successfully capped at 60FPS with the previous commit, the reason the player's animation played at an excessive speed was the fact that said animation is made up of only three frames, and changing the animation frame every 1/60th of a frame still resulted in an excessively fast animation.

This issue was solved by changing the animation frame every 8/60th of a second.

As I suspected, the extraction of a subsurface from the "playing_surface" surface was unnecessary for the correct displaying of the player's (frontal) animation; thus, it was removed.

A really rough player movement system was introduced: it is a directional key-triggered "pixel-based/perfect" movement (in the sense that - by keeping a directional key pressed, the player moves one pixel per game loop in the relative direction).

The final movement will have to be triggered by the angular velocity readings produced by the gyroscope and transmitted to the game by the board via Bluetooth; and it will probably be a "grid/tile-based" movement (in the sense that an angular velocity over a certain threshold along an axis - and with a certain verse - will cause the player to move towards a tile on the "pogo board" positioned UP, RIGHT, LEFT or RIGHT relative to the player's current position).