First refactoring of the code + markdown creation + "requirements.txt" put in place of "python_packages_to_install.txt"
The code up until the last commit worked just fine for its intended purpose, but it was all but modular.

Instead of implementing new logic, I've chosen to refactor the code so that it becomes a little more modular and scalable.

For now, I've written an abstract "Entity" class which models a generic object that can be positioned on the 26x15 grid of (24px,24px) tiles in which the playing surface is - at least conceptually - divided.

I've also written a "Player" class which inherits from the "Entity" class.

All of the player animation logic has been moved to said class, so that the variables and the instructions relative to the animation system do not clutter the "main program" (ledagogo.py) anymore.

In general, every object of an "Entity" subclass will have a "compute_surface" method which will get called for each and every one of them at the end of a "game loop" iteration: this method determines, at every game loop (and - according to the value of certain attributes of the "Entity" subclass object), what "surface" will represent the entity in the next game loop.

By doing so, the main program will just have to blit said surfaces on top of the correct surface ("playing_surface" or "game_surface"), and any change to the "graphical state" of the entity will be handled in a completely "internal" way (assumed/given that - obviously - the main program does not neglect calling the "compute_surface" method on/for every entity at the end of each game loop).

[N.B.]: The animation system for the player now accommodates for animations in all of the 4 possible directions of movement.

To test the functioning of said animation system, differently coloured frontal animation frames have been introduced in lieu of the actual animation frames for the remaining three directions.

The differently coloured frontal animations will be repurposed to be the frontal animation sprites for the other players, according to the following scheme:

[PLAYER COLOUR PROPOSAL]:
Player 1 -> Green
Player 2 -> Blue
Player 3 -> Red
Player 4 -> Yellow

(Which scheme has already been proposed on the "Frontal animation for the green player created" commit).