Multiple players on screen + initial implementation of a "collision handling system"
Given that the players' graphics are almost completed (only the back animation has yet to be done), I've tested the scenario in which multiple players (4 - to be exact) are present in the game session.

As of now, pressing the "TAB" key on the keyboard allows to select which player to control by using the directional keys: this is - obviously - solely for debugging purposes, given that each player's movements will be calculated on the basis of gyroscope readings acquired from different STM32F3DISCOVERY boards.

A rudimentary "collision handling system" has also been implemented.

In spite of its simplicity, it probably won't be changed, since it seems to work perfectly as intended.

It might be reworked in the case it reveals itself unfit for the "gyroscope-controlled" movement system.

One thing which could be changed is the player hitboxes: right now, they're 24x24 pixels "Rect" objects; and although they do the job just fine, they might be a little too big.

Reducing said hitboxes' sizes or changing their shape into something like a circle might be a better option...

But it probably won't be necessary to operate either of those changes, if we decide (and manage) to switch from a pixel-based player movement to a "grid/tile-based" player movement.