Game scenario graphics completely displayed
The game scenario graphics are now completely displayed in-game.

The only thing that is missing - graphics-wise - is the HUD.

Other nice variations would be having sparkles of the four colours of the players flashing on screen every once in a while, but I guess we'll do them only if we have enough time on our hands.