USART-based Bluetooth communication + LED turning on as the aftermath of a received command
Each LED gets turned on when the transmitter device sends a numeric code (corresponding to the LED) to the board.

The turning on of the LEDs is cumulative (that means, that if a "1" command is given, the LED corresponding to said numeric code gets turned on, but the remaining LEDs are not turned off).

As of now, the LEDs' order in turning on is not yet related to the game logic, as this commit is just an initial test.