Score calculation and displaying implemented
Every time a player closes a rectangle on the grid, the score gets updated by summing the player's currently held score with the area of the closed rectangle.

Displaying the score for each player on the HUD is now a functioning feature.

The HUD itself is still extremely barebones, both graphically (placeholder graphics are currently being used) and logic-wise (only the score updates get visualized on the HUD).

Updating the HUD's functioning will be the next task which will have to be fulfilled.

[N.B.]: The "find_smallest_rectangle" function has been modified to
accept a parameter which regulates the minimum size of the "smallest
rectangle" it has to detect (it previously detected rectangles which had
the minimum fixed size of (2,2)).

The new "smallest size" which has been chosen for the rectangle
detection is (3,3) [by doing so, we prevent the scenario in which a
player positions itself at the center of a 4-tile intersection and
starts going in circles to gather points].