"LED activation event" implemented (game-wise)
The game now detects if a player has exceeded a threshold associated with the activation ("turning on") of a LED on the board.

Obviously, for now the feature is only implemented in the game logic, but said event will also trigger the actual activation of a LED on the physical board when the project has reached its completion.