The new "PlayerScorer" graphics are properly displayed in-game
[TO DO]:
I'm not really that fond of the "PLAYER" label... I might have to colour it in a different fashion.

Also, the progress bar looks a little too dark, I'll probably have to lighten it up in some way.

The LEDs in the HUD look as I intended, but it would be nice to add a glowing effect to each of them (only when they're turned on, obviously).