First rough draft of the "power up" system
For now, the "PowerUp"s simply get spawned at random positions on the board.

We'll need to make them expire after a certain validity time (in seconds); and most of all, we'll need to make sure that a player "acquires" the power up if it happens to collide with one.

The acquired power up might be handled separately as a part of the "Player" object, but we'll evaluate this as we continue developing the "power up" system.