When a player acquires a power up, it gets displayed on the HUD
[N.B.]: the power up's validity does not get decremented when the reference to the PowerUp object gets assigned to the Player's object "power_up" attribute, so the power up can't expire yet when acquired.

Other than that, some tests have shown that the power up's validity remains at the value it had before the player's acquisition even after the player has acquired it.

A possible solution could be instantiating a new "PowerUp" object, and assigning the reference to this object to the Player's object's "power_up" attribute, but this seems to create a noticeable slow down in performances.

Thus, a simpler "data structure" could be ideal (the Player's object gets provided with a certain number of new attributes which model all the necessary information on the acquired power up, leaving other superfluous information - such as the PowerUp object's "grid_position" - unspecified).