A "player-acquired" power up is now able to expire
[N.B.]: when a player collides with a power up (and has already acquired one previously) the validity of the power up gets resetted to the initial value.

This is the desired effect, and it is achieved by simply re-executing the "power up acquisition" code.

What's left to do now - with regards to the power ups - is making it so that when a player holds a power up, said power up can be "consumed", and its consumption makes the player faster for a limited amount of time.